add rotationChange and shoot command handler.
fix player id always 0 issue.
add getRecentWorldState command and its handler.
add PlayerInfo class.
fix slot issue.

+++ AquaMan.Domain/Wallet.cs
@@ -5,7 +5,7 @@ namespace AquaMan.Domain
     public class Wallet
     {
         public Currency Currency { get; }
-        public ulong Amount { get; private set; }
+        public uint Amount { get; private set; }
         public uint Precise { get; }
 
         public Wallet(Currency currency, uint amount, uint precise)
